Split TypeScriptCompile into abstract and subclass.
